app/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
matplotlib
gunicorn
paho-mqtt